+++ CHANGELOG.md
@@ -13,6 +13,8 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 
 
 ## [Unreleased]
+### Added
+- ids parameter to transaction.query 
 
 ## [0.6.0] - 2020-07-12
 ### Added